% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfl1.R
\name{dfl1}
\alias{dfl1}
\title{Obtain degrees of freedom for l1ADMM objects}
\usage{
dfl1(tau, lambda, X, wNew, Z, S)
}
\arguments{
\item{tau}{vector of smoothing parameters for fixed effects}

\item{lambda}{scalar smoothing parameter for random effects}

\item{X}{list of fixed effect design matrices setup by l1smooth}

\item{wNew}{list of updated w parameters in ADMM algorithm}

\item{Z}{random effect design matrix setup by randomEffects}

\item{S}{random effect penalty matrix setup by randomEffects}
}
\description{
This function obtains degrees of freedom estimates
for models fit with the admm function. Usually called from within admm.
}
\examples{
dfTable <- dfl1(tau, lambda, X, wNew, Z, S)
}
\keyword{bands}
\keyword{confidence}
\keyword{intervals}

