% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funs.R
\name{Mfun}
\alias{Mfun}
\title{Function for making inverse of finite difference matrix}
\usage{
Mfun(k, p)
}
\arguments{
\item{k}{scalar k+1 is order of finite difference matrix}

\item{p}{scalar column dimension}
}
\description{
Function for making inverse of finite difference matrix
}
\keyword{Finite}
\keyword{difference}
\keyword{inverse}

