% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funs.R
\name{msub}
\alias{msub}
\title{sub function for making inverse of finite difference matrix}
\usage{
msub(i, p)
}
\description{
sub function for making inverse of finite difference matrix
}
\keyword{Finite}
\keyword{difference}
\keyword{inverse}

