% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{plot.ci}
\alias{plot.ci}
\title{Plot function for confidence bands}
\usage{
\method{plot}{ci}(CI)
}
\arguments{
\item{CI}{output from the ci function}
}
\description{
This function plots the confidence bands
}
\keyword{ci}
\keyword{print}

