% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{plot.cv}
\alias{plot.cv}
\title{Plot function for cv}
\usage{
\method{plot}{cv}(out)
}
\arguments{
\item{out}{output from the cv function}
}
\description{
This function plots the smoothing parameter paths 
and prints the optimal smoothing parameters
}
\keyword{cv}
\keyword{print}

