% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{print.ci}
\alias{print.ci}
\title{Print function for confidence bands}
\usage{
\method{print}{ci}(CI)
}
\arguments{
\item{CI}{output from the ci function}
}
\description{
This function plots the confidence bands
}
\keyword{ci}
\keyword{print}

