% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{print.cv}
\alias{print.cv}
\title{Print function for cv}
\usage{
\method{print}{cv}(out)
}
\arguments{
\item{out}{output from the cv function}
}
\description{
This function prints the optimal smoothing parameters
and prints the smoothing parameter paths
}
\keyword{cv}
\keyword{print}

