% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvgrid.R
\name{print.cvgrid}
\alias{print.cvgrid}
\title{Print function for cvgrid}
\usage{
\method{print}{cvgrid}(out)
}
\arguments{
\item{out}{output from the cvgrid function}
}
\description{
This function prints the optimal smoothing parameters
and prints the smoothing parameter paths
}
\keyword{cvgrid}
\keyword{print}

