% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrices.R
\name{ps}
\alias{ps}
\title{Function for setting up p-spline smooths}
\usage{
ps(x, norder, k, data, by = NULL, ref = NULL, width = NULL,
  delta = 1e-07, center = TRUE)
}
\arguments{
\item{x}{vector evaluation points (typically time)}

\item{norder}{scalar order of B-spline bases (order k is degree k-1)}

\item{k}{scalar k+1 is order of finite difference matrix}

\item{data}{dataset in which to evaluate arguments}

\item{by}{optional character string designating varying coefficient}

\item{ref}{optional reference level of by variable}

\item{width}{scalar width of B-spline bases}

\item{delta}{scalar place upper knot in B-spline basis at max(x) + delta}

\item{center}{TRUE/FALSE if True, impose centering constrain on non-variable coefficient models}
}
\description{
Function for setting up p-spline smooths
}
\examples{
library(psplinesl1)
data(simData)

# setup p-spline matrices
X <- list(ps(x = "x", data = simData, 
             norder = 2, k = 1, width = 0.05,
             center = TRUE))
}
\keyword{additive}
\keyword{components}

