% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrices.R
\name{psSub}
\alias{psSub}
\title{sub function for setting up p-spline smooths: called both by ps and ci}
\usage{
psSub(x, xcol, basis, norder, k, data, by = by, ref = ref, width = width,
  center = center)
}
\description{
sub function for setting up p-spline smooths: called both by ps and ci
}

