% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrices.R
\name{re}
\alias{re}
\title{Function for setting up random effects (intercepts or B-spline curves)}
\usage{
re(x, id, data, randomCurves = FALSE, width = 5, norder = 4,
  derivOrder = 2)
}
\arguments{
\item{x}{vector evaluation points (typically time)}

\item{id}{character string designating grouping variable}

\item{data}{dataset in which to evaluate arguments}

\item{randomCurves}{TRUE/FALSE if True, uses B-spline bases, otherwise random intercepts}

\item{width}{scalar width of B-spline bases (if randomCurves = TRUE)}

\item{norder}{scalar order of B-spline bases (order k is degree k-1) (if randomCurves = TRUE)}

\item{derivOrder}{scalar order of derivative used in the random effect penalty matrix (if randomCurves = TRUE)}
}
\description{
Function for setting up random effects (intercepts or B-spline curves)
}
\examples{
library(psplinesl1)
data(simData)

# setup random effect matrices
rand <- re(x = "x", id = "id", data = simData,
           randomCurves = FALSE)
}
\keyword{effects}
\keyword{random}

