% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfl1.R
\name{restricted}
\alias{restricted}
\title{This function returns the degrees of freedom using restricted derivatives}
\usage{
restricted(tau, Vj, Z, S)
}
\arguments{
\item{tau}{vector of smoothing parameters for random effects}
}
\description{
This function returns the degrees of freedom using restricted derivatives
}

