% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\docType{data}
\name{simData}
\alias{simData}
\title{Simulated data}
\format{A data frame with 450 rows of 3 variables:
\describe{
 \item{x}{covariate (could be time)}
 \item{y}{response}
 \item{id}{subject id (clustering variable)}
}}
\usage{
simData
}
\description{
A dataset containing simulated repeated measures data
}
\keyword{datasets}

