% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData2groups.R
\docType{data}
\name{simData2groups}
\alias{simData2groups}
\title{Simulated data}
\format{A data frame with 900 rows of 4 variables:
\describe{
 \item{x}{covariate (could be time)}
 \item{id}{subject id (clustering variable)}
 \item{y}{response}
 \item{group}{binary group indicator}
}}
\usage{
simData2groups
}
\description{
A dataset containing simulated repeated measures data
}
\keyword{datasets}

