% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funs.R
\name{soft}
\alias{soft}
\title{soft thresholding function}
\usage{
soft(kappa, x)
}
\arguments{
\item{kappa}{scalar threshold}

\item{x}{scalar value to be soft thresholded}
}
\description{
soft thresholding function
}
\keyword{soft}
\keyword{thresholding}

