% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfl1.R
\name{stein}
\alias{stein}
\title{This function returns the degrees of freedom using Steins' method}
\usage{
stein(tau, Vj, Z, S, activeSetCum, asLen, J)
}
\arguments{
\item{tau}{vector of smoothing parameters for random effects}
}
\description{
This function returns the degrees of freedom using Steins' method
}

