% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfl1.R
\name{tauRestricted}
\alias{tauRestricted}
\title{The root of this function gives tau}
\usage{
tauRestricted(tau, Vj, Z, S, sigma2b, SSE, sumn)
}
\arguments{
\item{tau}{vector of smoothing parameters for random effects}
}
\description{
The root of this function gives tau
}

