% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfl1.R
\name{tauStein}
\alias{tauStein}
\title{The root of this function gives tau}
\usage{
tauStein(tau, Vj, Z, S, activeSetCum, asLen, J, sigma2b, SSE, sumn)
}
\arguments{
\item{tau}{vector of smoothing parameters for random effects}
}
\description{
The root of this function gives tau
}

