% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueInteraction.R
\docType{data}
\name{trueInteraction}
\alias{trueInteraction}
\title{True marginal interaction for simulated data}
\format{A data frame with 101 rows of 3 variables:
\describe{
 \item{x}{covariate (could be time)}
 \item{y}{response}
 \item{id}{equal to 0 for all responses. Facilitates the plotting of trueMean with simData2groups}
}}
\usage{
trueInteraction
}
\description{
A dataset containing the true interaction used to generate simData2groups
}
\keyword{datasets}

