% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueMean.R
\docType{data}
\name{trueMean}
\alias{trueMean}
\title{True marginal mean for simulated data}
\format{A data frame with 101 rows of 3 variables:
\describe{
 \item{x}{covariate (could be time)}
 \item{y}{response}
 \item{id}{equal to 0 for all responses. Facilitates the plotting of trueMean with simData}
}}
\usage{
trueMean
}
\description{
A dataset containing the true marginal mean used to generate simData
}
\keyword{datasets}

